refactor session handling
